re-use instead of copy/paste
